Module iota_sc_utils::wasmlib

Integrated tool responsible for communication with ISCP's vm. This is a modified version of the original IF's wasmlib. Differences are kept minimal and only exist if they provide simpler interface, nothing else. Documentation: https://github.com/iotaledger/wasp/tree/master/contracts/rust/wasmlib/docs Github repository: https://github.com/brunoamancio/IOTA-SC-HName-Generator

Modules

host

keys

Structs

BytesDecoder

BytesEncoder

Key32

ScAddress

ScAgentId

ScBalances

ScChainId

ScColor

ScContractId

ScExports

ScFuncContext

ScHash

ScHname

ScImmutableAddress

ScImmutableAddressArray

ScImmutableAgentId

ScImmutableAgentIdArray

ScImmutableBytes

ScImmutableBytesArray

ScImmutableChainId

ScImmutableChainIdArray

ScImmutableColor

ScImmutableColorArray

ScImmutableContractId

ScImmutableContractIdArray

ScImmutableHash

ScImmutableHashArray

ScImmutableHname

ScImmutableHnameArray

ScImmutableInt64

ScImmutableInt64Array

ScImmutableMap

ScImmutableMapArray

ScImmutableRequestId

ScImmutableRequestIdArray

ScImmutableString

ScImmutableStringArray

ScMutableAddress

ScMutableAddressArray

ScMutableAgentId

ScMutableAgentIdArray

ScMutableBytes

ScMutableBytesArray

ScMutableChainId

ScMutableChainIdArray

ScMutableColor

ScMutableColorArray

ScMutableContractId

ScMutableContractIdArray

ScMutableHash

ScMutableHashArray

ScMutableHname

ScMutableHnameArray

ScMutableInt64

ScMutableInt64Array

ScMutableMap

ScMutableMapArray

ScMutableRequestId

ScMutableRequestIdArray

ScMutableString

ScMutableStringArray

ScRequestId

ScTransfers

ScUtility

ScViewContext

Constants

CORE_ACCOUNTS

CORE_ACCOUNTS_FUNC_DEPOSIT

CORE_ACCOUNTS_FUNC_WITHDRAW_TO_ADDRESS

CORE_ACCOUNTS_FUNC_WITHDRAW_TO_CHAIN

CORE_ACCOUNTS_PARAM_AGENT_ID

CORE_ACCOUNTS_VIEW_ACCOUNTS

CORE_ACCOUNTS_VIEW_BALANCE

CORE_ACCOUNTS_VIEW_TOTAL_ASSETS

CORE_BLOB

CORE_BLOB_FUNC_STORE_BLOB

CORE_BLOB_PARAM_FIELD

CORE_BLOB_PARAM_HASH

CORE_BLOB_VIEW_GET_BLOB_FIELD

CORE_BLOB_VIEW_GET_BLOB_INFO

CORE_BLOB_VIEW_LIST_BLOBS

CORE_EVENTLOG

CORE_EVENTLOG_PARAM_CONTRACT_HNAME

CORE_EVENTLOG_PARAM_FROM_TS

CORE_EVENTLOG_PARAM_MAX_LAST_RECORDS

CORE_EVENTLOG_PARAM_TO_TS

CORE_EVENTLOG_VIEW_GET_NUM_RECORDS

CORE_EVENTLOG_VIEW_GET_RECORDS

CORE_ROOT

CORE_ROOT_FUNC_CLAIM_CHAIN_OWNERSHIP

CORE_ROOT_FUNC_DELEGATE_CHAIN_OWNERSHIP

CORE_ROOT_FUNC_DEPLOY_CONTRACT

CORE_ROOT_FUNC_GRANT_DEPLOY_PERMISSION

CORE_ROOT_FUNC_REVOKE_DEPLOY_PERMISSION

CORE_ROOT_FUNC_SET_CONTRACT_FEE

CORE_ROOT_FUNC_SET_DEFAULT_FEE

CORE_ROOT_PARAM_CHAIN_OWNER

CORE_ROOT_PARAM_DEPLOYER

CORE_ROOT_PARAM_DESCRIPTION

CORE_ROOT_PARAM_HNAME

CORE_ROOT_PARAM_NAME

CORE_ROOT_PARAM_OWNER_FEE

CORE_ROOT_PARAM_PROGRAM_HASH

CORE_ROOT_PARAM_VALIDATOR_FEE

CORE_ROOT_VIEW_FIND_CONTRACT

CORE_ROOT_VIEW_GET_CHAIN_INFO

CORE_ROOT_VIEW_GET_FEE_INFO

KEY_ADDRESS

KEY_BALANCES

KEY_BASE58_BYTES

KEY_BASE58_STRING

KEY_BLS_ADDRESS

KEY_BLS_AGGREGATE

KEY_BLS_VALID

KEY_CALL

KEY_CALLER

KEY_CHAIN_OWNER_ID

KEY_COLOR

KEY_CONTRACT_CREATOR

KEY_CONTRACT_ID

KEY_DEPLOY

KEY_ED25519_ADDRESS

KEY_ED25519_VALID

KEY_EVENT

KEY_EXPORTS

KEY_HASH_BLAKE2B

KEY_HASH_SHA3

KEY_HNAME

KEY_INCOMING

KEY_LENGTH

KEY_LOG

KEY_MAPS

KEY_MINTED

KEY_NAME

KEY_PANIC

KEY_PARAMS

KEY_POST

KEY_RANDOM

KEY_REQUEST_ID

KEY_RESULTS

KEY_RETURN

KEY_STATE

KEY_TIMESTAMP

KEY_TRACE

KEY_TRANSFERS

KEY_UTILITY

KEY_VALID

KEY_ZZZZZZZ

Traits

MapKey

ScBaseContext