Crate iota_sc_utils

Modules

access
Responsible for handling access requirements. Fails calls on unauthorized access.

generator
Responsible for code generation. Integrated tool responsible for generating hashes in compile-time. Documentation: https://crates.io/crates/iota-sc-hname-generator Github repository: https://github.com/brunoamancio/IOTA-SC-HName-Generator

getter
Responsible for registering getter functions to contexts.

incoming
Syntax sugar to work with incoming transfers.

interfaces
Keeps information relevant to standards such as TIP-100.

math
Responsible for ensuring no under/overflows happen on maths calculations. Ensures calls panic on over/underflows.

params
Syntax sugar to work with context parameters.

results
Syntax sugar to work with context results.

state
Syntax sugar to work with context state.

wasmlib
Integrated tool responsible for communication with ISCP's vm. This is a modified version of the original IF's wasmlib. Differences are kept minimal and only exist if they provide simpler interface, nothing else. Documentation: https://github.com/iotaledger/wasp/tree/master/contracts/rust/wasmlib/docs Github repository: https://github.com/brunoamancio/IOTA-SC-HName-Generator