Module iota_sc_utils::wasmlib

Integrated tool responsible for communication with ISCP’s vm. This is a modified version of the original IF’s wasmlib. Differences are kept minimal and only exist if they provide simpler interface, nothing else. Documentation: <https://github.com/iotaledger/wasp/tree/master/contracts/rust/wasmlib/docs> Github repository: <https://github.com/brunoamancio/IOTA-SC-HName-Generator>

Modules

corecontracts

host

keys

Structs

BytesDecoder

BytesEncoder

Key32

ScAddress

ScAgentID

ScBalances

ScChainID

ScColor

ScExports

ScFunc

ScFuncContext

ScHash

ScHname

ScImmutableAddress

ScImmutableAddressArray

ScImmutableAgentID

ScImmutableAgentIDArray

ScImmutableBytes

ScImmutableBytesArray

ScImmutableChainID

ScImmutableChainIDArray

ScImmutableColor

ScImmutableColorArray

ScImmutableHash

ScImmutableHashArray

ScImmutableHname

ScImmutableHnameArray

ScImmutableInt16

ScImmutableInt16Array

ScImmutableInt32

ScImmutableInt32Array

ScImmutableInt64

ScImmutableInt64Array

ScImmutableMap

ScImmutableMapArray

ScImmutableRequestID

ScImmutableRequestIDArray

ScImmutableString

ScImmutableStringArray

ScInitFunc

ScMutableAddress

ScMutableAddressArray

ScMutableAgentID

ScMutableAgentIDArray

ScMutableBytes

ScMutableBytesArray

ScMutableChainID

ScMutableChainIDArray

ScMutableColor

ScMutableColorArray

ScMutableHash

ScMutableHashArray

ScMutableHname

ScMutableHnameArray

ScMutableInt16

ScMutableInt16Array

ScMutableInt32

ScMutableInt32Array

ScMutableInt64

ScMutableInt64Array

ScMutableMap

ScMutableMapArray

ScMutableRequestID

ScMutableRequestIDArray

ScMutableString

ScMutableStringArray

ScRequestID

ScTransfers

ScUtility

ScView

ScViewContext

Constants

KEY_ACCOUNT_ID

KEY_ADDRESS

KEY_BALANCES

KEY_BASE58_DECODE

KEY_BASE58_ENCODE

KEY_BLS_ADDRESS

KEY_BLS_AGGREGATE

KEY_BLS_VALID

KEY_CALL

KEY_CALLER

KEY_CHAIN_ID

KEY_CHAIN_OWNER_ID

KEY_COLOR

KEY_CONTRACT

KEY_CONTRACT_CREATOR

KEY_DEPLOY

KEY_ED25519_ADDRESS

KEY_ED25519_VALID

KEY_EVENT

KEY_EXPORTS

KEY_HASH_BLAKE2B

KEY_HASH_SHA3

KEY_HNAME

KEY_INCOMING

KEY_LENGTH

KEY_LOG

KEY_MAPS

KEY_MINTED

KEY_NAME

KEY_PANIC

KEY_PARAMS

KEY_POST

KEY_RANDOM

KEY_REQUEST_ID

KEY_RESULTS

KEY_RETURN

KEY_STATE

KEY_TIMESTAMP

KEY_TRACE

KEY_TRANSFERS

KEY_UTILITY

KEY_VALID

KEY_ZZZZZZZ

Traits

MapKey

ScBaseContext

ScFuncCallContext

ScViewCallContext

Functions

set_panic_hook